% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasSS-package.R
\docType{package}
\name{rasSS-package}
\alias{rasSS-package}
\alias{rasSS}
\title{rasSS: Summary-Statistic Extension of the Regional Association Score Method}
\description{
Extends the Regional Association Score (RAS) method of Jiang and Zhang (2025) to accept GWAS summary statistics. Replaces per-window individual-level regression with a 1-df weighted burden statistic derived from marginal Z-scores and an external LD reference, then runs the profile through the original change-point detection algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shan-team/RAS_GWAS_Summary}
  \item \url{https://shan-team.github.io/RAS_GWAS_Summary/}
  \item Report bugs at \url{https://github.com/shan-team/RAS_GWAS_Summary/issues}
}

}
\author{
\strong{Maintainer}: Anson Li \email{liyuanrui618@gmail.com}

Authors:
\itemize{
  \item Anson Li \email{liyuanrui618@gmail.com}
  \item Shengtong Han \email{shengtong.han@marquette.edu}
}

}
\keyword{internal}
