% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{Simulate genotypes from an LD matrix}
\usage{
sim_genotypes(n, R)
}
\arguments{
\item{n}{Integer. Number of individuals to simulate.}

\item{R}{Numeric matrix. LD correlation matrix.}
}
\value{
An \code{n} x \code{ncol(R)} integer matrix of simulated genotype dosages (values 0, 1, or 2).
}
\description{
Using a multivariate normal distribution constructed by the provided LD correlation matrix,
this function simulates genotypes in a population, then rounds and clips to
integer genotypic dosages (0, 1, 2).
}
